Use a unique name for CMake build-time-make-directory.
If CMake calls vulkan_compile_shader multiple times and the shader
directory does not exist in the build directory then
add_custom_target will be called multiple times with a target name of
build-time-make-directory which is not allowed.

Adding the filename of the input file to target name allows this to work
and fixes the build.

Signed-off-by: Robert Quill <robert.quill@imgtec.com>